v6.0 with more updates, just see

+++ modules/say.js
@@ -1,7 +1,11 @@
 module.exports = {
     desc: "Biarkan Aisha berbicara.",
     enable: true,
+    regex: false,
     role: ['433870492378595329'],
+    aliases: [],
+	usage: '[channel] [pesan anda]',
+	cooldown: 0,
     func: (client, message, args) => {
         const channel = message.mentions.channels.first();
         if (!channel) {
@@ -11,7 +15,11 @@ module.exports = {
             
         args.shift();
         let sayMessage = args.join(" ");
-                            
+        if (!sayMessage) {
+            message.channel.send("Mohon masukkan pesan anda!");
+            return;
+        }
+
         message.delete().catch(O_o=>{});
         channel.send(sayMessage);
     }
